fix android deps for gn check
