+++ android/BUILD.gn
@@ -8,7 +8,7 @@ declare_args() {
 
 java_cpp_enum("brave_android_java_enums_srcjar") {
   sources = [
-    "//brave/chromium_src/chrome/browser/notifications/notification_handler.h",
+    "//brave/chromium_src/chrome/browser/notifications/notification_handler_impl.h",
     "//brave/components/decentralized_dns/core/constants.h",
   ]
 }
