+++ android/android_browser_tests.gni
@@ -123,11 +123,9 @@ android_test_exception_deps = [
   "//brave/browser/net:browser_tests",
   "//brave/browser/permissions:browser_tests",
   "//brave/browser/test:browser_tests",
-  "//brave/browser/tor:browser_tests",
   "//brave/browser/ui:browser_tests",
   "//brave/browser/ui/tabs/test:browser_tests",
   "//brave/browser/ui/views/tabs:browser_tests",
-  "//brave/browser/widevine:browser_tests",
   "//brave/components/brave_wallet/resources:ethereum_provider_generated_resources",
   "//brave/components/de_amp/browser/test:browser_tests",
   "//brave/components/sessions/browser:browser_tests",
