+++ browser/net/brave_site_hacks_network_delegate_helper_browsertest.cc
@@ -8,7 +8,7 @@
 #include "base/strings/stringprintf.h"
 #include "brave/components/brave_shields/browser/brave_shields_util.h"
 #include "brave/components/constants/brave_paths.h"
-#include "brave/components/tor/onion_location_navigation_throttle.h"
+#include "brave/components/tor/buildflags/buildflags.h"
 #include "chrome/browser/content_settings/host_content_settings_map_factory.h"
 #include "chrome/browser/profiles/profile.h"
 #include "chrome/browser/ui/browser.h"
@@ -22,6 +22,10 @@
 #include "net/dns/mock_host_resolver.h"
 #include "net/test/embedded_test_server/default_handlers.h"
 #include "url/origin.h"
+
+#if BUILDFLAG(ENABLE_TOR)
+#include "brave/components/tor/onion_location_navigation_throttle.h"
+#endif
 
 class BraveSiteHacksNetworkDelegateBrowserTest : public InProcessBrowserTest {
  public:
