+++ browser/profiles/BUILD.gn
@@ -30,7 +30,7 @@ source_set("profiles") {
     "//brave/components/brave_wallet/common",
     "//brave/components/constants",
     "//brave/components/ipfs/buildflags",
-    "//brave/components/tor",
+    "//brave/components/tor/buildflags",
     "//chrome/common",
     "//components/gcm_driver:gcm_buildflags",
     "//components/keyed_service/content",
