+++ components/brave_wayback_machine/BUILD.gn
@@ -1,10 +1,6 @@
-import("//build/buildflag_header.gni")
-import("buildflags/buildflags.gni")
+import("//brave/components/brave_wayback_machine/buildflags/buildflags.gni")
 
-buildflag_header("buildflags") {
-  header = "buildflags.h"
-  flags = [ "ENABLE_BRAVE_WAYBACK_MACHINE=$enable_brave_wayback_machine" ]
-}
+assert(enable_brave_wayback_machine)
 
 static_library("brave_wayback_machine") {
   assert_no_deps = [ "//chrome/*" ]
@@ -37,6 +33,4 @@ static_library("brave_wayback_machine") {
     "//services/network/public/cpp",
     "//url",
   ]
-
-  public_deps = [ ":buildflags" ]
 }
